+++ CHANGELOG.md
@@ -13,7 +13,10 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Fix `ConversationsService::getServerConversations` to succeed even if the retrieve of informations about the contact failed.
 -   Add a `TimeOutManager` manager to be able to control every setTimeout from one tool. Useful for stopping setTimeout when a stop of SDK occured.
 -   Fix events treatment `ConversationHistoryHandler::onHistoryMessageReceived` when there are several times the tag `headers`. These events are raised as result of `conversations::getHistoryPage` API.
--   Add `AdminService::getAnImportStatus` API to provide a short status of the last import (completed or pending) of a company directory.    
+-   Add `AdminService::getAnImportStatus` API to provide a short status of the last import (completed or pending) of a company directory.
+-   Remove `PresenceService::enableCalendar` and `PresenceService::disableCalendar` because the API is not available on server side anymore.    
+-   Add `PresenceService::controlCalendarOrIgnoreAnEntry` API to Enable/disable a calendar sharing or ignore a calendar entry.
+-   Add `PresenceService::unregisterCalendar` API to Delete a calendar sharing.
     
 ## [2.17.0] - 2022-10-28
 -   Catch when getBubble failed.